+++ psm_utils/__init__.py
@@ -1,6 +1,7 @@
 """Common utilities for parsing and handling PSMs, and search engine results."""
 
-__version__ = "0.7.0"
+__version__ = "0.9.0"
+__all__ = ["Peptidoform", "PSM", "PSMList"]
 
 from warnings import filterwarnings
 
@@ -12,6 +13,6 @@ filterwarnings(
     module="psims.mzmlb",
 )
 
-from psm_utils.peptidoform import Peptidoform
-from psm_utils.psm import PSM
-from psm_utils.psm_list import PSMList
+from psm_utils.peptidoform import Peptidoform  # noqa: E402
+from psm_utils.psm import PSM  # noqa: E402
+from psm_utils.psm_list import PSMList  # noqa: E402